feat(Product): update product properties to use decimal types

+++ Inventory/Shared/Requests/UpdateProductRequest.cs
@@ -1,15 +1,14 @@
 namespace Shared.Requests;
 
-
 public class UpdateProductRequest
 {
     public string? Name { get; set; }
     public string? Description { get; set; }
     public decimal? UnitPrice { get; set; }
-    public float? Weight { get; set; }
-    public float? Length { get; set; }
-    public float? Width { get; set; }
-    public float? Height { get; set; }
+    public decimal? Weight { get; set; }
+    public decimal? Length { get; set; }
+    public decimal? Width { get; set; }
+    public decimal? Height { get; set; }
     public decimal? TaxCost { get; set; }
     public decimal? ProfitPerUnit { get; set; }
     public decimal? ProductionCost { get; set; }
